Ajusta casts de entrada em questao2.php

+++ src/questao2.php
@@ -1,5 +1,5 @@
 <?php
-$idade = readline("Digite sua idade: ");
+$idade = (int) readline("Digite sua idade: ");
 
 if ($idade < 12) {
     echo "Classificação: criança\n";
